+++ NEWS.md
@@ -5,6 +5,7 @@
 4. Made map optional in `read_midas_data`.
 5. Added `qp_genotypes` function.
 6. Added `mktest` function that works on allele table.
+7. Added `find_singletons` and `test_singleton_enrichment` functions.
 
 # 0.1-1
 1. Overall cleaning and NAMESPACE update